requirements.txt
discord.py==2.3.2
openai==1.3.7
python-dotenv==1.0.0

